Punto_5/requirements.txt
streamlit
transformers
diffusers
Pillow
torch
